+++ Src/Private/SharedUtilsFunctions.ps1
@@ -209,36 +209,3 @@ function ConvertTo-ADObjectName {
     }
     return $ADObject;
 }# end
-
-function ConvertTo-ADCanonicalName {
-    <#
-    .SYNOPSIS
-    Used by As Built Report to translate Active Directory DN to CanonicalName.
-    .DESCRIPTION
-
-    .NOTES
-        Version:        0.4.0
-        Author:         Jonathan Colon
-
-    .EXAMPLE
-
-    .LINK
-
-    #>
-    param(
-        [Parameter(Mandatory)]
-        [ValidateNotNullOrEmpty()]
-        $DN,
-        $Domain,
-        [pscredential]
-        $Credential
-    )
-    $ADObject = @()
-    $DC = Invoke-Command -Session $TempPssSession -ScriptBlock {Get-ADDomainController -Discover -Domain $using:Domain | Select-Object -ExpandProperty HostName}
-    $DCPssSession = New-PSSession $DC -Credential $Credential -Authentication Default
-    foreach ($Object in $DN) {
-        $ADObject += Invoke-Command -Session $DCPssSession {Get-ADObject $using:Object -Properties * | Select-Object -ExpandProperty CanonicalName}
-    }
-    Remove-PSSession -Session $DCPssSession
-    return $ADObject;
-}# end